+++ operators/__init__.py
@@ -36,7 +36,7 @@ from .meta_resize_to_content import POWER_SEQUENCER_OT_meta_resize_to_content
 from .meta_ungroup_and_trim import POWER_SEQUENCER_OT_meta_ungroup_and_trim
 from .meta_trim_content_to_bounds import POWER_SEQUENCER_OT_meta_trim_content_to_bounds
 from .mouse_trim_modal import POWER_SEQUENCER_OT_mouse_trim
-from .mouse_space_strips import POWER_SEQUENCER_OT_mouse_space_strips
+from .space_sequences import POWER_SEQUENCER_OT_space_sequences
 from .mouse_toggle_mute import POWER_SEQUENCER_OT_mouse_toggle_mute
 from .mouse_trim_instantly import POWER_SEQUENCER_OT_mouse_trim_instantly
 from .open_project_directory import POWER_SEQUENCER_OT_open_project_directory
@@ -105,7 +105,7 @@ classes = [
     POWER_SEQUENCER_OT_meta_ungroup_and_trim,
     POWER_SEQUENCER_OT_meta_trim_content_to_bounds,
     POWER_SEQUENCER_OT_mouse_trim,
-    POWER_SEQUENCER_OT_mouse_space_strips,
+    POWER_SEQUENCER_OT_space_sequences,
     POWER_SEQUENCER_OT_mouse_toggle_mute,
     POWER_SEQUENCER_OT_mouse_trim_instantly,
     POWER_SEQUENCER_OT_open_project_directory,